// app.client/src/api/models/IncomeTransactionListVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type IncomeTransactionListVM = {
    incomeTransactionID?: string | null;
    incomeId?: string | null;
    incomeTypeId?: string | null;
    name?: string | null;
    description?: string | null;
    dateOfIncome?: string;
    dateOfPayment?: string | null;
    incomeAmount?: number;
    customerId?: string | null;
    category?: string | null;
    categoryID?: string | null;
    isCleared?: boolean;
    isVoided?: boolean;
    createdBy?: string | null;
    createdDate?: string;
    lastModifiedBy?: string | null;
    lastModifiedDate?: string | null;
    incomeName?: string | null;
    incomeTypeName?: string | null;
    customerName?: string | null;
    isAdvance?: boolean;
    isDeposit?: boolean;
    depositReturnedDate?: string | null;
    depositReturnNotes?: string | null;
    mediaId?: string | null;
    mediaName?: string | null;
    budgetId?: string | null;
    budgetName?: string | null;
    incomeAmountApproved?: number;
    approvalId?: string | null;
    approvalName?: string | null;
    IsApproved?: boolean
};