// app.client/src/api/models/IncomeListVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type IncomeListVM = {
    incomeID?: string | null;
    incomeTypeId?: string | null;
    name?: string | null;
    description?: string | null;
    isVoided?: boolean;
    createdBy?: string | null;
    createdDate?: string;
    lastModifiedBy?: string | null;
    lastModifiedDate?: string | null;
    incomeName?: string | null;
    incomeTypeName?: string | null;
    incomeCategoryName?: string | null;
};